[TARGETING] Move writing of segment data to SegmentTracker
