+++ src/Targeting/ActionHandler/TrackSegment.php
@@ -17,26 +17,21 @@ declare(strict_types=1);
 
 namespace CustomerManagementFrameworkBundle\Targeting\ActionHandler;
 
-use CustomerManagementFrameworkBundle\Model\CustomerSegmentInterface;
+use CustomerManagementFrameworkBundle\Targeting\SegmentTracker;
 use Pimcore\Model\Tool\Targeting\Rule;
 use Pimcore\Targeting\ActionHandler\ActionHandlerInterface;
 use Pimcore\Targeting\Model\VisitorInfo;
-use Pimcore\Targeting\Session\SessionConfigurator;
-use Symfony\Component\HttpFoundation\Session\Attribute\NamespacedAttributeBag;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 class TrackSegment implements ActionHandlerInterface
 {
-    const KEY_SEGMENTS = 'cmf_segments';
-
     /**
-     * @var SessionInterface
+     * @var SegmentTracker
      */
-    private $session;
+    private $segmentTracker;
 
-    public function __construct(SessionInterface $session)
+    public function __construct(SegmentTracker $segmentTracker)
     {
-        $this->session = $session;
+        $this->segmentTracker = $segmentTracker;
     }
 
     /**
@@ -46,36 +41,5 @@ class TrackSegment implements ActionHandlerInterface
     {
         // TODO enable only if configured
         // TODO make configurable and assign configured segment
-    }
-
-    public function assignSegment(CustomerSegmentInterface $segment, int $count = 1)
-    {
-        $this->assignSegments([
-            $segment->getId() => $count
-        ]);
-    }
-
-    public function assignSegments(array $assignments)
-    {
-        $bag      = $this->getSessionBag();
-        $segments = $bag->get(self::KEY_SEGMENTS, []);
-
-        foreach ($assignments as $segmentId => $count) {
-            if (!isset($segments[$segmentId])) {
-                $segments[$segmentId] = 0;
-            }
-
-            $segments[$segmentId] += $count;
-        }
-
-        $bag->set(self::KEY_SEGMENTS, $segments);
-    }
-
-    private function getSessionBag(): NamespacedAttributeBag
-    {
-        /** @var NamespacedAttributeBag $bag */
-        $bag = $this->session->getBag(SessionConfigurator::TARGETING_BAG);
-
-        return $bag;
     }
 }
